# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Gear Road Roulette — game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface CharacterStats {
  /** 1-10 scale */
  confidence: number;
  /** 1-10 scale */
  mechanical: number;
  /** 1-10 scale */
  navigation: number;
  /** starting funds in pounds */
  budget: number;
}

export interface Character {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  personality: string;
  stats: CharacterStats;
}

export type MissionDifficulty = typeof MissionDifficulty[keyof typeof MissionDifficulty];


export const MissionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  insane: 'insane',
} as const;

export interface Mission {
  id: number;
  title: string;
  location: string;
  terrain: string;
  description: string;
  budget: number;
  difficulty: MissionDifficulty;
}

export type MissionDetailDifficulty = typeof MissionDetailDifficulty[keyof typeof MissionDetailDifficulty];


export const MissionDetailDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  insane: 'insane',
} as const;

export interface CarOption {
  id: number;
  name: string;
  year: number;
  price: number;
  /** 1-10 */
  reliability: number;
  /** 1-10 */
  power: number;
  /** 1-10 */
  offRoad: number;
  description: string;
}

export type ChallengeType = typeof ChallengeType[keyof typeof ChallengeType];


export const ChallengeType = {
  race: 'race',
  skill: 'skill',
  survival: 'survival',
  beauty: 'beauty',
  drag: 'drag',
} as const;

export interface Challenge {
  id: number;
  title: string;
  type: ChallengeType;
  description: string;
}

export interface MissionDetail {
  id: number;
  title: string;
  location: string;
  terrain: string;
  description: string;
  budget: number;
  difficulty: MissionDetailDifficulty;
  availableCars: CarOption[];
  challenges: Challenge[];
}

export type GameSaveStatus = typeof GameSaveStatus[keyof typeof GameSaveStatus];


export const GameSaveStatus = {
  car_selection: 'car_selection',
  on_road: 'on_road',
  challenge: 'challenge',
  completed: 'completed',
  failed: 'failed',
} as const;

export type GameSaveMode = typeof GameSaveMode[keyof typeof GameSaveMode];


export const GameSaveMode = {
  arcade: 'arcade',
  series: 'series',
} as const;

export interface GameSave {
  id: number;
  /** @nullable */
  characterId: number | null;
  missionId: number;
  status: GameSaveStatus;
  mode: GameSaveMode;
  /** @nullable */
  playerName: string | null;
  seriesStageIndex: number;
  funds: number;
  /** @nullable */
  carId: number | null;
  food: number;
  parts: number;
  camaraderie: number;
  distanceTravelled: number;
  score: number;
  createdAt: string;
  updatedAt: string;
}

export type GameSaveInputMode = typeof GameSaveInputMode[keyof typeof GameSaveInputMode];


export const GameSaveInputMode = {
  arcade: 'arcade',
  series: 'series',
} as const;

export interface GameSaveInput {
  /** @nullable */
  characterId?: number | null;
  missionId: number;
  mode?: GameSaveInputMode;
  /** @nullable */
  playerName?: string | null;
  seriesStageIndex?: number;
}

export type GameSaveUpdateStatus = typeof GameSaveUpdateStatus[keyof typeof GameSaveUpdateStatus];


export const GameSaveUpdateStatus = {
  car_selection: 'car_selection',
  on_road: 'on_road',
  challenge: 'challenge',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface GameSaveUpdate {
  status?: GameSaveUpdateStatus;
  missionId?: number;
  funds?: number;
  /** @nullable */
  carId?: number | null;
  seriesStageIndex?: number;
  food?: number;
  parts?: number;
  camaraderie?: number;
  distanceTravelled?: number;
  score?: number;
}

export type RoadEventEventType = typeof RoadEventEventType[keyof typeof RoadEventEventType];


export const RoadEventEventType = {
  breakdown: 'breakdown',
  weather: 'weather',
  police: 'police',
  shortcut: 'shortcut',
  fuel: 'fuel',
  mechanical: 'mechanical',
  wildlife: 'wildlife',
  banter: 'banter',
} as const;

export interface RoadEvent {
  id: number;
  saveId: number;
  eventType: RoadEventEventType;
  title: string;
  description: string;
  outcome: string;
  fundsChange: number;
  createdAt: string;
}

export type RoadEventInputEventType = typeof RoadEventInputEventType[keyof typeof RoadEventInputEventType];


export const RoadEventInputEventType = {
  breakdown: 'breakdown',
  weather: 'weather',
  police: 'police',
  shortcut: 'shortcut',
  fuel: 'fuel',
  mechanical: 'mechanical',
  wildlife: 'wildlife',
  banter: 'banter',
} as const;

export interface RoadEventInput {
  eventType: RoadEventInputEventType;
  title: string;
  description: string;
  outcome: string;
  fundsChange: number;
}

export interface LeaderboardEntry {
  id: number;
  /** @nullable */
  saveId?: number | null;
  playerName: string;
  characterSlug: string;
  missionTitle: string;
  score: number;
  distance: number;
  createdAt: string;
}

export interface LeaderboardInput {
  /** @nullable */
  saveId?: number | null;
  playerName: string;
  characterSlug: string;
  missionTitle: string;
  score: number;
  distance?: number;
}

export type BanterRequestTone = typeof BanterRequestTone[keyof typeof BanterRequestTone];


export const BanterRequestTone = {
  excited: 'excited',
  sarcastic: 'sarcastic',
  horrified: 'horrified',
  smug: 'smug',
  panicked: 'panicked',
  competitive: 'competitive',
} as const;

export interface BanterRequest {
  /** Current game situation/event description */
  context: string;
  /** Active character slugs e.g. ['jeremy', 'richard', 'james'] */
  characters: string[];
  tone?: BanterRequestTone;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListLeaderboardParams = {
limit?: number;
};