# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Top Gear Road Roulette — game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [character]
      summary: Get a single character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Missions ─────────────────────────────────────────────────────────────────
...
  /missions/{id}:
    get:
      operationId: getMission
      tags: [mission]
      summary: Get a single mission with available cars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Game saves ────────────────────────────────────────────────────────────────
...
  /leaderboard:
    get:
      operationId: listLeaderboard
      tags: [leaderboard]
      summary: Get top leaderboard entries (ordered by score desc)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
    post:
      operationId: createLeaderboardEntry
      tags: [leaderboard]
      summary: Submit a score to the global leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderboardInput"
      responses:
        "201":
          description: Created leaderboard entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"

  # ── LLM Banter ───────────────────────────────────────────────────────────────
  /banter/generate:
    post:
      operationId: generateBanter
      tags: [banter]
      summary: Generate in-character banter from the three hosts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanterRequest"
      responses:
        "200":
          description: SSE stream of banter lines
          content:
            text/event-stream: {}

  # ── OpenAI endpoints ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Characters ──────────────────────────────────────────────────────────────
    Character:
      type: object
      required: [id, slug, name, tagline, personality, stats]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        tagline:
          type: string
        personality:
          type: string
        stats:
          $ref: "#/components/schemas/CharacterStats"
    CharacterStats:
      type: object
      required: [confidence, mechanical, navigation, budget]
      properties:
        confidence:
          type: integer
          description: "1-10 scale"
        mechanical:
          type: integer
          description: "1-10 scale"
        navigation:
          type: integer
          description: "1-10 scale"
        budget:
          type: integer
          description: "starting funds in pounds"

    # ── Missions ─────────────────────────────────────────────────────────────────
...
    MissionDetail:
      type: object
      required: [id, title, location, terrain, description, budget, difficulty, availableCars, challenges]
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        terrain:
          type: string
        description:
          type: string
        budget:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, insane]
        availableCars:
          type: array
          items:
            $ref: "#/components/schemas/CarOption"
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/Challenge"
    CarOption:
      type: object
      required: [id, name, year, price, reliability, power, offRoad, description]
      properties:
        id:
          type: integer
        name:
          type: string
        year:
          type: integer
        price:
          type: integer
        reliability:
          type: integer
          description: "1-10"
        power:
          type: integer
          description: "1-10"
        offRoad:
          type: integer
          description: "1-10"
        description:
          type: string
    Challenge:
      type: object
      required: [id, title, type, description]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [race, skill, survival, beauty, drag]
        description:
          type: string

    # ── Game Save ──────────────────────────────────────────────────────────────
...
    LeaderboardEntry:
      type: object
      required: [id, playerName, characterSlug, missionTitle, score, distance, createdAt]
      properties:
        id:
          type: integer
        saveId:
          type: ["integer", "null"]
        playerName:
          type: string
        characterSlug:
          type: string
        missionTitle:
          type: string
        score:
          type: integer
        distance:
          type: integer
        createdAt:
          type: string
          format: date-time
    LeaderboardInput:
      type: object
      required: [playerName, characterSlug, missionTitle, score]
      properties:
        saveId:
          type: ["integer", "null"]
        playerName:
          type: string
        characterSlug:
          type: string
        missionTitle:
          type: string
        score:
          type: integer
        distance:
          type: integer

    # ── Banter ────────────────────────────────────────────────────────────────
    BanterRequest:
      type: object
      required: [context, characters]
      properties:
        context:
          type: string
          description: "Current game situation/event description"
        characters:
          type: array
          items:
            type: string
          description: "Active character slugs e.g. ['jeremy', 'richard', 'james']"
        tone:
          type: string
          enum: [excited, sarcastic, horrified, smug, panicked, competitive]

    # ── OpenAI ────────────────────────────────────────────────────────────────
...