# lib/api-zod/src/generated/types/carOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Gear Road Roulette — game API
 * OpenAPI spec version: 0.1.0
 */

export interface CarOption {
  id: number;
  name: string;
  year: number;
  price: number;
  /** 1-10 */
  reliability: number;
  /** 1-10 */
  power: number;
  /** 1-10 */
  offRoad: number;
  description: string;
}