# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Gear Road Roulette — game API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeType } from './challengeType';

export interface Challenge {
  id: number;
  title: string;
  type: ChallengeType;
  description: string;
}