# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Gear Road Roulette — game API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterStats } from './characterStats';

export interface Character {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  personality: string;
  stats: CharacterStats;
}